AIHelper/requirements.txt
pathfinding==0.0.4
pyastar==0.0.1
scikit_image==0.18.1
numpy==1.18.4
Django==3.1.4
matplotlib==3.3.3
networkx==2.5
cheroot==8.5.1
scipy==1.5.3
numba==0.52.0
bjoern==3.1.0
Pillow==8.1.0
skimage==0.0
